Assistant Builder: use hook for template management
